+++ .github/workflows/verify.yml
@@ -62,7 +62,7 @@ jobs:
     timeout-minutes: 5
     strategy:
       matrix:
-        typescript-version: ['5.0.4', '5.1.6', '5.2.2', 'latest']
+        typescript-version: ['5.0.4', '5.1.6', '5.2.2', '5.3.3', '5.4.5']
         viem-version: ['2.9.31', 'latest']
 
     steps:
