+++ packages/vue/src/composables/useWriteContract.test-d.ts
@@ -1,7 +1,6 @@
-import { http, type WriteContractErrorType, createConfig } from '@wagmi/core'
-import { base } from '@wagmi/core/chains'
+import { type WriteContractErrorType } from '@wagmi/core'
 import { abi } from '@wagmi/test'
-import { type Abi, type Address, type Hash, parseEther } from 'viem'
+import { type Abi, type Address, type Hash } from 'viem'
 import { expectTypeOf, test } from 'vitest'
 
 import { useSimulateContract } from './useSimulateContract.js'
@@ -141,71 +140,4 @@ test('useSimulateContract', () => {
 
   const request = data?.value?.request
   if (request) writeContract(request)
-})
-
-// https://github.com/wevm/wagmi/issues/3981
-test('gh#3981', () => {
-  const config = createConfig({
-    chains: [base],
-    transports: {
-      [base.id]: http(),
-    },
-  })
-  const { writeContract } = useWriteContract({ config })
-
-  const abi = [
-    {
-      type: 'function',
-      name: 'example1',
-      inputs: [
-        { name: 'exampleName', type: 'address', internalType: 'address' },
-      ],
-      outputs: [],
-      stateMutability: 'payable',
-    },
-    {
-      type: 'function',
-      name: 'example2',
-      inputs: [
-        { name: 'exampleName', type: 'address', internalType: 'address' },
-      ],
-      outputs: [],
-      stateMutability: 'nonpayable',
-    },
-  ] as const
-
-  writeContract({
-    abi,
-    address: '0x...',
-    functionName: 'example1',
-    args: ['0x...'],
-    value: parseEther('1'),
-  })
-
-  writeContract({
-    abi: [
-      {
-        type: 'function',
-        name: 'example1',
-        inputs: [
-          { name: 'exampleName', type: 'address', internalType: 'address' },
-        ],
-        outputs: [],
-        stateMutability: 'payable',
-      },
-      {
-        type: 'function',
-        name: 'example2',
-        inputs: [
-          { name: 'exampleName', type: 'address', internalType: 'address' },
-        ],
-        outputs: [],
-        stateMutability: 'nonpayable',
-      },
-    ],
-    address: '0x...',
-    functionName: 'example1',
-    args: ['0x...'],
-    value: parseEther('1'),
-  })
 })
